# perf(treesitter): don't fetch parser for each line