@justinmk 2025-01-02
Is it worth grabbing the test from https://github.com/neovim/neovim/pull/27422 ?