@MariaSolOs 2025-01-07
For my own learning: Are injections relevant here? Or is only the root parser needed for `foldexpr`?